First basic push - Few rough elements
This is very much a work in progress. I got really tired halfway through developing this. The water background is likely not going to stay, but I thought the white was boring so I did something about it.
Pretty much all that is final here is the text within the div. Div styles and getting that image to cooperate will propbably happen in my next push.